+++ extension/content-script.js
@@ -1,10 +1,73 @@
 const STORAGE_KEY = "darkModeSettings";
 const OPTIONS = {
-  DARK: "invert(1) hue-rotate(180deg)",
-  SEPIA: "sepia(1)",
-  GRAY_SCALE: "grayscale(1)",
-  DISABLED: "none",
+  DARK: applyDarkModeStyles,
+  SEPIA: applySepiaModeStyles,
+  GRAY_SCALE: applyGrayScaleModeStyles,
+  DISABLED: applyDisabledModeStyles,
 };
+
+function applyDisabledModeStyles() {
+  const style = document.createElement("style");
+  style.innerHTML = `
+    body {
+      filter: none;
+      background-color: inherit;
+    }
+
+    *::selection {
+      background: #ee0;
+      color: #222;
+    }
+  `;
+  document.head.appendChild(style);
+}
+
+function applyGrayScaleModeStyles() {
+  const style = document.createElement("style");
+  style.innerHTML = `
+    body {
+      filter: grayscale(1);
+    }
+
+    *::selection {
+      background: #8e8e0f;
+      color: #000;
+    }
+  `;
+  document.head.appendChild(style);
+}
+
+function applySepiaModeStyles() {
+  const style = document.createElement("style");
+  style.innerHTML = `
+    body {
+      filter: sepia(1);
+    }
+
+    *::selection {
+      background: #8e8e0f;
+      color: #000;
+    }
+  `;
+  document.head.appendChild(style);
+}
+
+function applyDarkModeStyles() {
+  const style = document.createElement("style");
+  style.innerHTML = `
+    body {
+      filter: invert(1) hue-rotate(180deg);
+      background-color: #000;
+    }
+
+    *::selection {
+      background: #8e8e0f;
+      color: #000;
+    }
+  `;
+
+  document.head.appendChild(style);
+}
 
 /**
  * Add event listener to the chrome runtime message
@@ -13,8 +76,11 @@ chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
   const { option, hostname } = request;
 
   if (option in OPTIONS) {
-    document.body.style.filter = OPTIONS[option];
-    document.body.style.backgroundColor = option === "DARK" ? "#000" : "";
+    const applyStyles = OPTIONS[option];
+
+    if (typeof applyStyles === "function") {
+      applyStyles();
+    }
 
     if (hostname) {
       saveSettings(hostname, option);