+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2023-12-18
+
+### Added
+
+- Refactored code to extend styles according to mode selected, instead of only using a filter.
+
+## [1.0.0] - 2023-12-10
+
+### Added
+
+- First official release of Dark Mode Enabler Chrome Extension.
+
 ## [0.0.2] - 2023-12-06
 
 ### Added